[ANDROSDK-1821] Do not include page if paging false (old exporter)
