+++ core/src/main/java/org/hisp/dhis/android/core/trackedentity/internal/TrackedEntityInstanceService.kt
@@ -75,8 +75,8 @@ internal interface TrackedEntityInstanceService {
         @Query(FIELDS) @Which fields: Fields<TrackedEntityInstance>,
         @Query(ORDER) order: String?,
         @Query(PAGING) paging: Boolean,
-        @Query(PAGE) page: Int,
-        @Query(PAGE_SIZE) pageSize: Int,
+        @Query(PAGE) page: Int?,
+        @Query(PAGE_SIZE) pageSize: Int?,
         @Query(LAST_UPDATED_START_DATE) lastUpdatedStartDate: String?,
         @Query(INCLUDE_ALL_ATTRIBUTES) includeAllAttributes: Boolean,
         @Query(INCLUDE_DELETED) includeDeleted: Boolean,
@@ -105,8 +105,8 @@ internal interface TrackedEntityInstanceService {
         @Query(LAST_UPDATED_END_DATE) lastUpdatedEndDate: String?,
         @Query(ORDER) order: String?,
         @Query(PAGING) paging: Boolean,
-        @Query(PAGE) page: Int,
-        @Query(PAGE_SIZE) pageSize: Int,
+        @Query(PAGE) page: Int?,
+        @Query(PAGE_SIZE) pageSize: Int?,
     ): SearchGrid
 
     companion object {
